Move ThingyMock to PCLMock directory+namespace
