+++ Benchmarks/EmptyMethodOnly.cs
@@ -1,4 +1,5 @@
 using BenchmarkDotNet.Attributes;
+using BenchmarkMockNet.PCLMock;
 using FakeItEasy;
 using Moq;
 using NSubstitute;
